Petite réorganisation de la main branche (et j'en profite pour tester comment marcher les commit depuis pycharm)
